CaféInsight BI
This Business Intelligence (BI) project delves into the analysis of a one-month dataset from a café, encompassing around 40,000 records. The visualization focuses on key aspects such as Sales Trends, Customer Persona Analysis, and Pastry Inventory Analysis.

The project aims to provide comprehensive insights to the café, aiding in operational improvement, cost reduction, and the optimization of the marketing strategy.